==> 5_Data_Cleaning.ipynb <==
# Import Packages

import os
import pandas as pd

# Load Match Data

# Load the CSV file into a DataFrame
matches_csv_path = "matches_2.csv"
matches_df = pd.read_csv(matches_csv_path)

# Display the DataFrame
matches_df.head()

matches_df.shape

# Data Cleaning

matches_df["home_player_X1"]

def process_match_data(df):
    """
    Preprocess the matches DataFrame by performing the following steps:
    1. Drop specific in-game event columns and rank-related columns.
    2. Drop columns with more than 400 missing values.
    3. Drop rows with any missing values.
    4. Remove rows with invalid formations for home and away teams.

    Input: df (pd.DataFrame): The input DataFrame to preprocess.

    Output: pd.DataFrame: The preprocessed DataFrame.
    """

    # Drop player position columns since we already have frmation and do not need them
    for i in range(11):

      df.drop(columns = [f"home_player_X{i+1}"], inplace = True)
      df.drop(columns = [f"away_player_X{i+1}"], inplace = True)
      df.drop(columns = [f"home_player_Y{i+1}"], inplace = True)
      df.drop(columns = [f"away_player_Y{i+1}"], inplace = True)

    # add Year and Month columns since we will drop date
    df["date"] = pd.to_datetime(df["date"])
    df["Year"] = df["date"].dt.year
    df["Month"] = df["date"].dt.month

    # Drop in-game events and rank-related columns
    columns_to_drop = [
        "id", "home_team_goal", "away_team_goal", "date", "goal_difference", "date"
    ]
    df.drop(columns=columns_to_drop, inplace=True, errors='ignore')

    # Drop columns with more than 400 missing values
    df.drop(columns=[col for col in df.columns if df[col].isnull().sum() > 400], inplace=True)

    # Drop rows with any missing values
    df.dropna(inplace=True)

    # Define invalid formations for home and away teams
    invalid_home_formations = ['1-1-1-1-1-1-1-1-1-1', '6-3', '4-2-2-1']
    invalid_away_formations = ['1-1-1-1-1-1-1-1-1-1', '2-3-1-1', '3-3-1-1-1', '3-2-1-1']

    # Remove rows with invalid formations
    df = df[~df["home_team_formation"].isin(invalid_home_formations)]
    df = df[~df["away_team_formation"].isin(invalid_away_formations)]

    return df

# Apply the function to matches_df
matches_df = process_match_data(matches_df)

# Display the result
matches_df

# Save to CSV

# Create the directory
output_dir = "Data"
if not os.path.exists(output_dir):
    os.makedirs(output_dir)

# Save the matches_df DataFrame to a CSV file
output_path = os.path.join(output_dir, "matches_3.csv")
matches_df.to_csv(output_path, index=False)

print(f"matches_df has been saved to {output_path}")
